\name{annotatedDataFrameFrom-methods}
\docType{methods}
\alias{annotatedDataFrameFrom}
\alias{annotatedDataFrameFrom,AssayData-method}

\title{Methods for Function annotatedDataFrameFrom in Package `Biobase'}
\description{
  \code{annotatedDataFrameFrom} is a convenience for creating
  \code{\link{AnnotatedDataFrame}} objects.
}
\section{Methods}{
\describe{
  \item{\code{annotatedDataFrameFrom(assayData, byrow=FALSE)}}{This
    method creates an \code{AnnotatedDataFrame} using sample (when
    \code{byrow=FALSE}) or feature (\code{byrow=TRUE}) names and dimensions
    of an \code{\link{AssayData}} object as a template.}
}} 

\author{Biocore team}

\keyword{methods} 
