\name{Biobase-package}
\alias{Biobase-package}
\alias{Biobase}
\docType{package}
\title{Biobase Package Overview}
\description{
   Biobase Package Overview
}
\details{
   Main Classes to operate with data: \code{\link{ExpressionSet}}, \code{\link{MultiSet}}, \code{\link{eSet}}, , \code{\link{exprSet}}, \code{\link{MIAME}}, \code{\link{phenoData}}. Full help on methods and associated functions is available from withing class help pages.

   Additional manipulation and data structuring classes: \code{\link{Versioned}}, \code{\link{VersionedBiobase}}, \code{\link{aggregator}}, \code{\link{container}}.

   Vignette routines: \code{\link{openVignette}}, \code{\link{getPkgVigs}}, \code{\link{openPDF}}.

   Package manipulation functions: \code{\link{createPackage}} and \code{\link{package.version}}

   Data sets: \code{\link{aaMap}}, \code{\link{sample.ExpressionSet}}, \code{\link{sample.exprSet}}, \code{\link{sample.eSet}}, \code{\link{geneData}}.

   Introductory information is available from vignettes, type \code{openVignette()}.

   Full listing of documented articles is available in HTML view by typing \code{help.start()} and selecting \code{Biobase} package from the \code{Packages} menu or via \code{library(help="Biobase")}.
}
\author{
   O. Sklyar
}
\keyword{package}
