\name{characterORMIAME-class}
\docType{class}
\alias{characterORMIAME-class}
\alias{characterORMIAME}
\title{Class characterORMIAME, a class to make older versions compatible}
\description{This class can be either \code{character} or
  \code{MIAME}. It is for the description slot in exprSet}
\section{Methods}{
No methods defined with class "characterORMIAME" in the signature.
}
\seealso{ See also \code{\link{MIAME-class}}, \code{\link{exprSet-class}}}
\keyword{classes}
