\name{class:characterORMIAME}
\docType{class}
\alias{class:characterORMIAME}
\concept{characterORMIAME}
\title{Class to Make Older Versions Compatible}
\description{This class can be either \code{character} or
  \code{MIAME}. It is for the description slot in exprSet}
\section{Methods}{
No methods defined with class "characterORMIAME" in the signature.
}
\seealso{ See also \code{\link{MIAME}}, \code{\link{exprSet}}}
\keyword{classes}
