\name{class:exprMatrix}
\docType{class}
\alias{class:exprMatrix}
\alias{exprMatrix-class}

\concept{exprMatrix}
\concept{expression matrix}

\title{Class Union for Objects Behaving Like a Matrix }
\description{  Class union 'exprMatrix' represents the type of \code{exprs} and
  \code{se.exprs} slots in 'exprSet' objects. One of the classes in the union
  is always "matrix" }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "exprMatrix" in the signature.
}
\keyword{classes}
