\name{data:eset}
\alias{data:eset}
\alias{eset}
\non_function{}
\title{(Deprecated) dataset of class 'exprSet'}
\usage{data(eset)}
\format{
The data for 26 cases, labeled A to Z and 500 genes.
}
\description{
 This name is deprecated. Use \code{\link{sample.exprSet.1}} instead.

 The data are real but anonymized. The data are from an experiment that used Affymetrix
 U95v2 chips. The data were processed by dChip and then exported to R
 for analysis.
}
\source{
 Sabina Chiaretti of the Ritz Laboratory at the Dana Farber Cancer
 Institute.
}
\examples{
data(sample.exprSet.1)
}
\keyword{datasets}
