\name{data:sample.exprSet.1}
\alias{data:sample.exprSet.1}
\alias{sample.exprSet.1}
\non_function{}
\title{Dataset of class 'exprSet'}
\usage{data(sample.exprSet.1)}
\format{
The data for 26 cases, labeled A to Z and 500 genes.
}
\description{
 The data are real but anonymized. The data are from an experiment that used Affymetrix
 U95v2 chips. The data were processed by dChip and then exported to R
 for analysis. The phenoData component is made up. 
}
\source{
 Sabina Chiaretti of the Ritz Laboratory at the Dana Farber Cancer
 Institute.
}
\examples{
data(sample.exprSet.1)
}
\keyword{datasets}
