\name{Deprecated}
\title{Biobase Deprecated and Defunct}

\alias{addVarMetadataEntry}
\alias{addVarMetadataEntry,phenoData,character,character-method}

\alias{getUnits}
\alias{getUnits,phenoData,character-method}

\alias{eList}
\alias{eList<-}
\alias{eList<-,eSet,AssayData-method}
\alias{eList,eSet-method}
\alias{eList<-,eSet,AssayData-method}

\description{
   The function you have asked for has been deprected.  It may even be
   defunct.
}
\keyword{internal}


