\name{eset}
\alias{eset}
\non_function{}
\title{ A dataset of class exprSet  }
\usage{data(eset)}
\format{
 The data are real but anonymized. There are 26 cases, labeled A to Z
 and 500 genes. The data are from an experiment that used Affymetrix
 U95v2 chips. The data were processed by dChip and then exported to R
 for analysis.
}
\description{
  The data are described above.
  }
\source{
 Sabina Chiaretti of the Ritz Laboratory at the Dana Farber Cancer
 Institute. 
}
\examples{
data(eset)
}
\keyword{datasets}
