\name{exprMatrix-class}
\docType{class}
\alias{exprMatrix-class}

\title{ Class union for objects behaving like a matrix }
\description{  Class union "exprMatrix" represents the type of exprs and
  se.exprs slots in "exprSet" objects. One of the classes in the union
  is always "matrix" }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "exprMatrix" in the signature.
}
\keyword{classes}
