\name{abstract}
%generics
\alias{abstract}
\alias{addNonExisting}
\alias{addPDF2Vig}
\alias{addVarMetadataEntry}
\alias{addVig2Menu}
\alias{addVig4Unix}
\alias{addVig4Win}
\alias{aggenv}
\alias{aggfun}
\alias{annotation}
\alias{annotation<-}
\alias{assayData<-}
\alias{content}
\alias{convertVarLabels}
\alias{description}
\alias{description<-}
\alias{df2pD}
\alias{eList<-}
\alias{expinfo}
\alias{exprs}
\alias{exprs<-}
\alias{geneNames}
\alias{geneNames<-}
\alias{getExpData}
\alias{getUnits}
\alias{getVarMetadata}
\alias{hybridizations}
\alias{initfun}
\alias{iter}
\alias{locked}
\alias{normControls}
\alias{notes}
\alias{notes<-}
\alias{otherInfo}
\alias{pData<-}
\alias{phenoData<-}
\alias{preproc}
\alias{read.pD}
\alias{reporterInfo}
\alias{reporterInfo<-}
\alias{reporterNames}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{samples}
\alias{se.exprs}
\alias{se.exprs<-}
\alias{varLabels}

%classes: this notation is irritating...
\alias{MIAME-class}
\alias{eSet-class}
\alias{exprSet-class}
\alias{aggregator-class}
\alias{annotatedDataset-class}
\alias{characterORMIAME-class}
\alias{container-class}
\alias{data.frameOrNULL-class}
\alias{exprMatrix-class}
\alias{listOrEnv-class}
\alias{phenoData-class}

%datasets
\alias{bbsym}
\alias{geneCov}

%methods
\alias{abstract,MIAME-method}
\alias{abstract,exprSet-method}
\alias{addVarMetadataEntry,phenoData,character,character-method}
\alias{aggenv,aggregator-method}
\alias{aggfun,aggregator-method}
\alias{annotation<-,eSet-method}
\alias{annotation<-,exprSet-method}
\alias{annotation,eSet-method}
\alias{annotation,exprSet-method}
\alias{as.data.frame.exprSet,exprSet-method}
\alias{assayData<-,eSet,listOrEnv-method}
\alias{assayData,eSet-method}
\alias{assayData,exprSet-method}
\alias{[,container-method}
\alias{[,phenoData-method}
\alias{[,eSet-method}
\alias{[,exprSet-method}
\alias{[[<-,container-method}
\alias{[[<-,phenoData-method}
\alias{[[<-,annotatedDataset-method}
\alias{[[,container-method}
\alias{[[,phenoData-method}
\alias{[[,annotatedDataset-method}
\alias{$<-,phenoData-method}
\alias{$<-,annotatedDataset-method}
\alias{$,phenoData-method}
\alias{$,annotatedDataset-method}
\alias{coerce,data.frame,phenoData-method}
\alias{coerce,exprSet,eSet-method}
\alias{combine,phenoData,phenoData-method}
\alias{combine,phenoData,ANY-method}
\alias{combine,eSet,eSet-method}
\alias{combine,eSet,ANY-method}
\alias{content,container-method}
\alias{contents,environment-method}
\alias{convertVarLabels,phenoData-method}
\alias{description<-,eSet-method}
\alias{description<-,exprSet-method}
\alias{description,eSet-method}
\alias{description,exprSet-method}
\alias{dim,eSet-method}
\alias{eList<-,eSet,listOrEnv-method}
\alias{eList,eSet-method}
\alias{expinfo,MIAME-method}
\alias{exprs2excel,exprSet-method}
\alias{exprs<-,eSet,listOrEnv-method}
\alias{exprs<-,exprSet-method}
\alias{exprs,eSet-method}
\alias{exprs,exprSet-method}
\alias{geneNames<-,exprSet-method}
\alias{geneNames,exprSet-method}
\alias{getExpData,eSet,character-method}
\alias{getUnits,phenoData,character-method}
\alias{getVarMetadata,phenoData,character,character-method}
\alias{getVarMetadata,phenoData,character,missing-method}
\alias{hybridizations,MIAME-method}
\alias{initfun,aggregator-method}
\alias{iter,exprSet,missing,function-method}
\alias{iter,exprSet,missing,list-method}
\alias{iter,exprSet,character,function-method}
\alias{length,container-method}
\alias{locked,container-method}
\alias{ncol,eSet-method}
\alias{normControls,MIAME-method}
\alias{notes<-,eSet-method}
\alias{notes<-,exprSet-method}
\alias{notes,eSet-method}
\alias{notes,exprSet-method}
\alias{otherInfo,MIAME-method}
\alias{pData<-,phenoData-method}
\alias{pData<-,annotatedDataset-method}
\alias{pData,phenoData-method}
\alias{pData,annotatedDataset-method}
\alias{phenoData<-,annotatedDataset,phenoData-method}
\alias{phenoData,annotatedDataset-method}
\alias{preproc,MIAME-method}
\alias{reporterInfo<-,annotatedDataset,data.frame-method}
\alias{reporterInfo,annotatedDataset-method}
\alias{reporterNames,eSet-method}
\alias{sampleNames<-,eSet,character-method}
\alias{sampleNames<-,exprSet-method}
\alias{sampleNames,eSet-method}
\alias{sampleNames,exprSet-method}
\alias{samples,MIAME-method}
\alias{se.exprs<-,exprSet-method}
\alias{se.exprs,exprSet-method}
\alias{show,container-method}
\alias{show,phenoData-method}
\alias{show,MIAME-method}
\alias{show,eSet-method}
\alias{show,exprSet-method}
\alias{split,exprSet,factor,missing-method}
\alias{split,exprSet,vector,missing-method}
\alias{split,phenoData,vector,missing-method}
\alias{update2MIAME,exprSet-method}
\alias{varLabels,phenoData-method}
\alias{varLabels,annotatedDataset-method}
\alias{varMetadata,phenoData-method}
\alias{write.exprs,exprSet-method}

\title{Internals}
\description{
   use help.search("your keyword")
}
\keyword{internal}
