\name{selectSome}
\alias{selectSome}
\title{extract aspects of a vector for concise rendering  }
\description{
extract aspects of a vector for concise rendering  
}
\usage{
selectSome(obj, maxToShow=5)
}
\arguments{
  \item{obj}{ vector }
  \item{maxToShow}{ number of elements to render}
}
%\details{
%}
\value{
string coercion with ellipsis intercalated


}
%\references{  }
\author{Biocore}
%\note{  }


%\seealso{  }

\examples{
selectSome(1:20)
}
\keyword{ models }
