\name{characterORmiame-class}
\docType{class}
\alias{characterORmiame-class}
\alias{characterORmiame}
\title{Class characterORmiame, a class to make older versions compatible}
\description{This class can be either \code{character} or
  \code{miame}. It is for the description slot in exprSet}
\section{Methods}{
No methods defined with class "characterORmiame" in the signature.
}
\seealso{ See also \code{\link{miame-class}}, \code{\link{exprSet-class}}}
\keyword{classes}
