\name{class:characterORMIAME}
\docType{class}
\alias{class:characterORMIAME}
\alias{characterORMIAME-class}

\concept{characterORMIAME}
\title{Class to Make Older Versions Compatible}
\description{This class can be either \code{character} or
  \code{MIAME}.}
\section{Methods}{
No methods defined with class "characterORMIAME" in the signature.
}
\seealso{See also \code{\link{MIAME}}}
\keyword{classes}
