\name{data:geneData}
\alias{data:geneData}
\alias{geneData}
\concept{geneCov}
\non_function{}
\title{ A subset of some expression data. }
\usage{data(geneData)}
\description{
The \code{geneData} data frame has 500 rows and 26 columns.
It consists of a subset of real expression data from an Affymetrix U95v2
chip. The data are anonymous. The covariate data \code{geneCov} is made
up.
}
\format{
  A 500 by 26 data frame.
}
\source{
  The J. Ritz Laboratory (S. Chiaretti).
}

\examples{
data(geneData)
}
\keyword{datasets}
