\name{Deprecated and Defunct}
\title{Biobase Deprecated and Defunct}

\alias{class:annotatedDataset}
\alias{annotatedDataset}
\alias{annotatedDataset-class}
\alias{initialize,annotatedDataset-method}

\alias{class:exprList}
\alias{exprList}
\alias{exprList-class}

\alias{class:exprMatrix}
\alias{exprMatrix}
\alias{exprMatrix-class}

\alias{class:exprSet}
\alias{exprSet}
\alias{exprSet-class}
\alias{initialize,exprSet-method}

\alias{class:phenoData}
\alias{phenoData-class}
\alias{initialize,phenoData-method}

\alias{data:sample.eSet}
\alias{sample.eSet}

\alias{data:sample.exprSet.1}
\alias{sample.exprSet.1}

\alias{data:sample.exprSet}
\alias{sample.exprSet}

\alias{df2pD}
\alias{read.pD}
\alias{read.exprSet}
\alias{read.phenoData}

\alias{geneNames}
\alias{geneNames,ExpressionSet-method}
\alias{geneNames<-}
\alias{geneNames<-,ExpressionSet,character-method}

\alias{getExpData}
\alias{getExpData,eSet,character-method}

\alias{eList}
\alias{eList,eSet-method}
\alias{eList<-}
\alias{eList<-,eSet,AssayData-method}

\alias{reporterNames}
\alias{reporterNames,eSet-method}
\alias{reporterNames<-}
\alias{reporterNames<-,eSet,character-method}

\alias{updateOldMiame}

\alias{l2e}

\alias{biocReposList}

\description{
   The function, class, or data object you have asked for has been deprecated or made defunct. 
}

\keyword{internal}
