\name{abstract}
%generics
\alias{abstract}
\alias{addNonExisting}
\alias{addPDF2Vig}
\alias{addVarMetadataEntry}
\alias{addVig2Menu}
\alias{addVig4Unix}
\alias{addVig4Win}
\alias{aggenv}
\alias{aggfun}
\alias{content}
\alias{convertVarLabels}
\alias{description}
\alias{description<-}
\alias{df2pD}
\alias{eList<-}
\alias{expinfo}
\alias{getExpData}
\alias{getUnits}
\alias{hybridizations}
\alias{initfun}
\alias{iter}
\alias{locked}
\alias{normControls}
\alias{otherInfo}
\alias{preproc}
\alias{read.pD}
\alias{reporterInfo}
\alias{reporterInfo<-}
\alias{samples}
\alias{se.exprs}
\alias{se.exprs<-}
\alias{notes}
\alias{notes<-}


%classes: this notation is irritating...
\alias{data.frameOrNULL-class}
\alias{listOrEnv-class}

%datasets
\alias{bbsym}
\alias{geneCov}

%methods
\alias{contents,environment-method}

\title{Internals}
\description{
   Use help.search("your keyword", package="Biobase").
}
\keyword{internal}
