\name{Internals}
\alias{aggenv}
\alias{aggfun}
\alias{content}
\alias{expinfo}
\alias{hybridizations}
\alias{initfun}
\alias{locked}
\alias{normControls}
\alias{otherInfo}


%classes
\alias{data.frameOrNULL-class}
\alias{listOrEnv-class}

%methods
\alias{contents,environment-method}

\title{Internals}
\description{
   Use help.search("your keyword", package="Biobase").
}
\keyword{internal}
